Added form functionality into the new flow + minor controller and logic changes across program

+++ app/Http/Controllers/LeadController.php
@@ -9,7 +9,7 @@ use App\Http\Requests\UpdateLeadRequest;
 use Illuminate\Support\Facades\Storage;
 use App\Http\Resources\LeadResource;
 use Illuminate\Support\Facades\File;
-use Illuminate\Http\Client\Request;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Http\UploadedFile;
 use App\Models\Lead;
@@ -29,10 +29,10 @@ class LeadController extends Controller
     public function store(StoreLeadsRequest $request)
     {
         $data = $request->validated();
-        $request->user()->lead()
-        ? $lead = $request->user()->lead()
+        $request->user()->lead
+        ? $lead = $request->user()->lead
         : $lead = Lead::where('email', $data['email'])->first();
-        $isCompleteLead = $lead->status === 'complete';
+        $isCompleteLead = $lead ? $lead->status == 'complete' : false;
         
         if ($isCompleteLead) {
             return response('Submission already completed.', 409);
@@ -46,14 +46,25 @@ class LeadController extends Controller
         return new LeadResource($lead);
     }
 
+    public function getStatus(Request $request) {
+        $user = $request->user();
+        return response()->json([
+            'leadStatus' => $user->lead->status ?? 'empty'
+        ]);
+    }
+
     public function show(Request $request, Lead $lead)
     {
-        if ($lead->status == 'failed') {
+        if ($lead && $lead->status == 'failed') {
             File::delete(storage_path($lead->proof_of_id));
             File::delete(storage_path($lead->proof_of_address));
             $lead->proof_of_id = null;
             $lead->proof_of_address = null;
             $lead->save();
+        } else if(!$lead) {
+            return response()->json([
+                'message' => 'no lead found'
+            ], 404);
         }
         return new LeadResource($lead);
     }
@@ -96,7 +107,7 @@ class LeadController extends Controller
         } return null;
     }
     private function storeImageForLead(Lead $lead, UploadedFile $file, string $kind) {
-        $dir = "images/leads/{$lead->id}";
+        $dir = "images/leads/{$lead->user_id}";
         $ext = strtolower($file->guessExtension() ? : 'bin');
         $name = "{$kind}" . ".{$ext}";
